% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_battery.R
\name{add_bbg_ticker}
\alias{add_bbg_ticker}
\title{Load BBG ticker to list for data load}
\usage{
add_bbg_ticker(bbg_ticker)
}
\arguments{
\item{bbg_ticker}{BBG Ticker to include}
}
\value{
Include BBG Ticker to table for data load
}
\description{
Load BBG ticker to list for data load
}
\examples{
\donttest{add_bbg_ticker("AAPL US Equity")}
}
