% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat_active_risk.R
\name{bdh_weekday}
\alias{bdh_weekday}
\title{Wrapper for bdh with pre-built options for getting daily data for all weekdays, non-trading weekdays will have previous value}
\usage{
bdh_weekday(series, field = "PX_LAST",
  start_date = as.Date("1994-01-01"), end_date = NULL,
  options_overrides = NULL)
}
\arguments{
\item{field}{A character vector with Bloomberg query fields.}

\item{start_date}{A Date variable with the query start date.}

\item{end_date}{An optional Date variable with the query end date; if omitted the most recent available date is used.}

\item{securities}{A character vector with security symbols in Bloomberg notation.}

\item{options}{An optional named character vector with additional option values}
}
\value{
A list with as a many entries as there are entries in securities; each list contains a data.frame with one row per observations and as many columns as entries in fields. If the list is of length one, it is collapsed into a single data frame. Note that the order of securities returned is determined by the backend and may be different from the order of securities in the securities field.
}
\description{
Wrapper for bdh with pre-built options for getting daily data for all weekdays, non-trading weekdays will have previous value
}
\examples{
\donttest{
bdh(c("SPX Index", "STI Index"), "PX_LAST", start.date = Sys.Date() - 31)
}
}
