% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat_active_risk.R
\name{calc_active_risk}
\alias{calc_active_risk}
\title{Calculate active risk given return of strategies and weight of the strategies}
\usage{
calc_active_risk(unwt_ret, curr_wt, start_date = today() - years(10),
  end_date = today(), annualize_factor = 250)
}
\arguments{
\item{unwt_ret}{dataframe of returns of strategies with columns `date`, `strategy`, `return`}

\item{curr_wt}{dataframe of weights with columns `strategy`, `size`}

\item{start_date}{start_date for calc of active risk}

\item{end_date}{end_date}

\item{annualize_factor}{factor to multiply by to convert to annual. 250 for daily data (default) and 12 for monthly data}
}
\value{
list containing `cov_matrix`, `port_sd` (portfolio std dev), `active_risk`, `marginal_risk`
}
\description{
Calculate active risk given return of strategies and weight of the strategies
}
\examples{
portfolios <- build_strategies(demo_strategies,
  start_date = as.Date("2016-01-01"),
  end_date = as.Date("2018-12-07"))
sim_pf_size <- convert_dur_size(portfolios$sim, portfolios$summary, demo_duration)
wt_return <- calc_strat_wt_return(sim_pf_size, demo_return)
headline_size <- calc_strat_headline_size(sim_pf_size)
unwt_return <- calc_strat_unwt_return(wt_return, headline_size)
curr_wt <- get_strat_size(headline_size, as.Date("2018-01-05"))
calc_active_risk(unwt_return, curr_wt, as.Date("2017-06-01"), as.Date("2018-06-01"))
}
