% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat_active_risk.R
\name{calc_cor}
\alias{calc_cor}
\title{Calculate correlation between returns}
\usage{
calc_cor(unwt_ret_df, start_date = NA, end_date = NA,
  period_name = NULL)
}
\arguments{
\item{unwt_ret_df}{dataframe containing timeseries returns for each strategy}

\item{start_date}{A Date variable which is the start_date of the period for correlation calculation}

\item{end_date}{A Date variable}

\item{period_name}{optional period_name to attach to the returned correlations}
}
\value{
long form dataframe (see tidyr::gather) containing correlation between returns, and `period_name`` if provided
}
\description{
Calculate correlation between returns
}
\examples{
unwt_ret <- data.frame(date = as.Date(c("2018-01-02", "2018-01-02", "2018-01-04", "2018-01-05")),
                              long_spx = c(0.015, 0.021, -0.03, 0.01),
                              long_ukx = c(-0.005, 0.03, -0.01, -0.04),
                              long_hsi = c(0.023, 0.001, -0.005, 0.008))
calc_cor(unwt_ret)
}
