% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat_active_risk.R
\name{calc_returns}
\alias{calc_returns}
\title{Calculate individual daily returns of a dataframe containing multiple assets}
\usage{
calc_returns(df)
}
\arguments{
\item{df}{dataframe containing asset prices, `date` column is optional}
}
\value{
dataframe containing returns
}
\description{
Calculate individual daily returns of a dataframe containing multiple assets
}
\examples{
df <- data.frame(spx = c(2850, 2820, 2890), ukx = c(7004, 7010, 7080))
calc_returns(df)
}
