% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttool.R
\name{calc_sharpe}
\alias{calc_sharpe}
\title{Calculate Sharpe}
\usage{
calc_sharpe(ret, annualize_factor = 252)
}
\arguments{
\item{ret}{return}

\item{annualize_factor}{Factor to use for annualise provided return}
}
\value{
Sharpe ratio
}
\description{
Calculate Sharpe
}
\examples{
calc_sharpe(0.1)
}
