% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat_active_risk.R
\name{calc_strat_headline_size}
\alias{calc_strat_headline_size}
\title{Calculates headline sizes of strategies given the size of the component trades by first summing the positive sizes and negative sizes separately. Then taking the higher number as the size}
\usage{
calc_strat_headline_size(strat_list)
}
\arguments{
\item{strat_list}{A list of strategies which are dataframes containing timeseries sizes of their component instruments}
}
\value{
A dataframe containing the size of the strategies daily
}
\description{
Calculates headline sizes of strategies given the size of the component trades by first summing the positive sizes and negative sizes separately. Then taking the higher number as the size
}
