% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat_active_risk.R
\name{calc_strat_wt_return}
\alias{calc_strat_wt_return}
\title{Calculate weighted return of strategies}
\usage{
calc_strat_wt_return(strat_list, asset_returns)
}
\arguments{
\item{strat_list}{A list containing dataframes for each strategy with their corresponding timeseries sizes in percent or month-weighted sizes. Column name must be the instrument's identifier}

\item{asset_returns}{A dataframe with time-series returns of all instruments with column name as the `identifier`}
}
\value{
A dataframe containing the time-series daily return of all instruments
}
\description{
Calculate weighted return of strategies
}
\examples{
\donttest{
portfolios <- build_strategies(gen_strat_templates())
dur <- get_dur_bbg(portfolios$summary)
actual_pf_size <- convert_dur_size(portfolios$actual, portfolios$summary, dur)
ret <- get_ret_bbg(portfolios$summary)
calc_strat_wt_return(actual_pf_size, ret)
}
}
