% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat.R
\name{clean_rating}
\alias{clean_rating}
\title{Clean rating}
\usage{
clean_rating(rating)
}
\arguments{
\item{rating}{Bloomberg rating returned}
}
\value{
Standardised Bloomberg rating used internally
}
\description{
Clean rating
}
