% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat_active_risk.R
\name{custom_grouping}
\alias{custom_grouping}
\title{Group returns of strategies}
\usage{
custom_grouping(returns_df, instr_df, group)
}
\arguments{
\item{returns_df}{A dataframe containing timeseries of returns of strategies with columns `date`, `strategy`, `return`}

\item{instr_df}{A dataframe containing `strategy` corresponding to any groupings}

\item{group}{A character variable containing the name of the column in `instr_df` to group by}
}
\value{
A dataframe containing timeseries of returns of the groups
}
\description{
Group returns of strategies
}
