% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autosend.R
\name{email_update}
\alias{email_update}
\title{Distribute daily trading update}
\usage{
email_update()
}
\value{
Side effect only: sends email to distribution list
}
\description{
Distribute daily trading update
}
\examples{
\donttest{email_update()}
}
