% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot.R
\name{en.pivot}
\alias{en.pivot}
\title{Function for Pivot Points}
\usage{
en.pivot(tseries)
}
\arguments{
\item{tseries}{dataframe contain columns date, open, high, low, close}
}
\value{
dataframe containing pivots
}
\description{
Function for Pivot Points
}
\examples{
library(dplyr)
data(swat_data)
df_rx <- swat_data \%>\% dplyr::filter(Ticker == "RX1 Comdty") \%>\% dplyr::select(-Ticker)
en.pivot(df_rx)
}
