% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqd_rebal.R
\name{eqd}
\alias{eqd}
\title{Asset class rebalancing}
\usage{
eqd(port.SGD, weight, units)
}
\arguments{
\item{port.SGD}{Size of portfolio in Singapore dollars}

\item{weight}{percentage allocation to asset class}

\item{units}{Existing units of asset class}
}
\value{
Number of units to add to portfolio (-ve means reduce allocation)
}
\description{
Asset class rebalancing
}
\examples{
\donttest{
eqd(1000000, 0.05, 10)
}
}
