% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat_active_risk.R
\name{get_and_check_tickers}
\alias{get_and_check_tickers}
\title{Get tickers of available securities in the specific asset class, and check which securities in the instruments dataframe are valid for the asset class}
\usage{
get_and_check_tickers(instruments_df, type = c("price", "duration"))
}
\arguments{
\item{instruments_df}{A dataframe containing the instruments to search the tickers for. Dataframe must have the `asset_class` and `identifier` columns}

\item{type}{A character which can be `price` or `duration`, which indicates whether to return tickers for price or duration. Defaults to `price`}
}
\value{
A dataframe with the matched instruments from instruments_df and their corresponding tickers
}
\description{
Get tickers of available securities in the specific asset class, and check which securities in the instruments dataframe are valid for the asset class
}
\examples{
\donttest{
inst <- data.frame(asset_cl = c("equity", "ilb", "govt"),
                    identifier = c("spx", "germany_ilb_5y", "us_govt_10y"),
                    stringsAsFactors = FALSE)
get_and_check_tickers("govt", inst)
}
}
