% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{get_bbg_data}
\alias{get_bbg_data}
\title{Cache Bloomberg data locally}
\usage{
get_bbg_data(ticker, start.date = Sys.Date() - 365 * 1,
  end.date = Sys.Date())
}
\arguments{
\item{ticker}{Single ticker to save OHLC and volume data}

\item{start.date}{Start date}

\item{end.date}{End date}
}
\value{
Saves a csv file in the data2 folder
}
\description{
Cache Bloomberg data locally
}
\examples{
\donttest{
Rblpapi::blpConnect()
get_bbg_data("AAPL US Equity")
}
}
