% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat.R
\name{get_bm_ratings}
\alias{get_bm_ratings}
\title{Download and calculate median credit rating over time}
\usage{
get_bm_ratings(end_date = Sys.Date(), per = 120)
}
\arguments{
\item{end_date}{Last date to show credit ratings.}

\item{per}{No. of historical monthends to retrieve}
}
\value{
Returns a data frame to be used with plot_credit_ratings()
}
\description{
Download and calculate median credit rating over time
}
\examples{
\donttest{get_bm_ratings(start_date="2018-10-01", end_date = "2018-11-01")}
}
