% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat_active_risk.R
\name{get_dur_bbg}
\alias{get_dur_bbg}
\title{Get time-series duration of asset from Bloomberg. It is a generic function which means new duration functions can be easily added for new assets}
\usage{
get_dur_bbg(instruments_df, start_date = as.Date("1994-01-01"),
  end_date = today(), fill = TRUE)
}
\arguments{
\item{instruments_df}{A dataframe containing the columns `identifier` and `asset_class`}

\item{start_date}{A Date variable}

\item{end_date}{A Date variable}

\item{fill}{A boolean indicating whether NAs in the data should be filled, data is only filled upwards}
}
\value{
A dataframe containing the the time-series duration of the assets, with the identifier as the name
}
\description{
Get time-series duration of asset from Bloomberg. It is a generic function which means new duration functions can be easily added for new assets
}
\examples{
\donttest{
inst <- data.frame(asset_class = c("ilb", "govt", "fut", "cds"),
   identifier = c("germany_ilb_5y", "us_govt_10y", "fut_rx1", "cdx_us_ig"),
   stringsAsFactors = FALSE)
get_dur_bbg(inst)
}
}
