% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat.R
\name{get_mas_rating}
\alias{get_mas_rating}
\title{Get MAS rating}
\usage{
get_mas_rating(my_sec, date)
}
\arguments{
\item{my_sec}{A list of Bloomberg tickers to calculate ratings}

\item{date}{Date to obtain rating}
}
\value{
Clean ratings and MAS rating
}
\description{
Get MAS rating
}
\examples{
## library(Rblpapi)
## library(lubridate)
## blpConnect()
## get_mas_rating("GTGBP10Y Govt", ymd("20180101"))

}
