% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat_active_risk.R
\name{get_ret_bbg}
\alias{get_ret_bbg}
\title{Get time-series daily returns of asset from Bloomberg. It is a generic function which means new return functions can be easily added for new assets}
\usage{
get_ret_bbg(instruments_df, start_date = as.Date("1994-01-01"),
  end_date = today())
}
\arguments{
\item{instruments_df}{A dataframe containing the columns `asset_class` and `identifier`}

\item{start_date}{A Date variable}

\item{end_date}{A Date variable}
}
\value{
A gathered dataframe containing timeseries daily returns with columns `date`, `instrument` and `return`
}
\description{
Get time-series daily returns of asset from Bloomberg. It is a generic function which means new return functions can be easily added for new assets
}
\examples{
\donttest{
inst <- data.frame(asset_class = c("ilb", "govt", "fut", "cds"),
    identifier = c("germany_ilb_5y", "us_govt_10y", "fut_rx1", "cdx_us_ig"),
    stringsAsFactors = FALSE)
get_ret_bbg(inst)
}
}
