% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat_active_risk.R
\name{get_strat_size}
\alias{get_strat_size}
\title{Get strategies sizes at a specific date from headline sizes of strategies}
\usage{
get_strat_size(strat_df, as_of_date = NULL, approx = TRUE)
}
\arguments{
\item{strat_df}{A dataframe containing timeseries of headline sizes of strategies}

\item{as_of_date}{A Date variable}

\item{approx}{A boolean indicating if `as_of_date` is not available, the next available earlier date will be used}
}
\value{
A dataframe containing the columns `strat` and `size`
}
\description{
Get strategies sizes at a specific date from headline sizes of strategies
}
