% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat_active_risk.R
\name{get_tickers}
\alias{get_tickers}
\title{Get tickers for the specific asset class from the package's csv files}
\usage{
get_tickers(asset_cl, roll_differencing = TRUE)
}
\arguments{
\item{asset_cl}{A character variable of `govt`, `fut`, `equity`, `cds`, or `fx`}

\item{roll_differencing}{A boolean indicating if ticker for futures should be modified to account for roll_differencing, defaults as TRUE}
}
\value{
A dataframe containing the tickers and their corresponding identifiers
}
\description{
Get tickers for the specific asset class from the package's csv files
}
\examples{
get_tickers("fx")
}
