% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{getData}
\alias{getData}
\title{Get data from Bloomberg using the Rblpapi package}
\usage{
getData(Ticker)
}
\arguments{
\item{ticker}{A character string containing a Bloomberg Ticker to download}

\item{start_date}{Start date to retrieve data. Defaults to one year ago.}

\item{end_date}{End date ti retrieve data. Defaults to today.}

\item{fields}{A list of Bloomberg fields to download. Defaults to PX_LAST}
}
\value{
A data frame containing the fields of provided ticker for a given period of time.
}
\description{
Get data from Bloomberg using the Rblpapi package
}
\examples{
getData("AAPL US Equity")
}
