% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{getData_xts}
\alias{getData_xts}
\title{Get price time series from Bloomberg, using the Rblpapi package}
\usage{
getData_xts(ticker, start_date = Sys.Date() - 365, end_date = today())
}
\arguments{
\item{ticker}{A character string containing a Bloomberg Ticker to download}

\item{start_date}{Start date to retrieve data. Defaults to one year ago.}

\item{end_date}{End date ti retrieve data. Defaults to today.}
}
\value{
A xts time series containing OHLC data
}
\description{
Get price time series from Bloomberg, using the Rblpapi package
}
\examples{
getData_xts("AAPL US Equity")
}
