% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{ggTS}
\alias{ggTS}
\title{Draws a simple time series plot}
\usage{
ggTS(Ticker, Title = Ticker)
}
\arguments{
\item{ticker}{Character string containing Bloomberg ticker}

\item{title}{Character string to use in chart title}

\item{yield_mode}{Boolean flag to indicate if subtitle should indicate 1-day yield changes instead of 1-day percentage returns}

\item{start_date}{Start date for chart}
}
\description{
Draws a simple time series plot
}
\examples{
ggTS("AAPL US Equity")
}
