% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_battery.R
\name{load_db}
\alias{load_db}
\title{read BBG data stored in db}
\usage{
load_db(dbname = "inst/extdata/mydatabase.db")
}
\arguments{
\item{dbname}{Path of sql}
}
\value{
Dataframe containing data in px_db
}
\description{
read BBG data stored in db
}
\examples{
\donttest{load_db()}
}
