% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat.R
\name{market_capping}
\alias{market_capping}
\title{Market_capping}
\usage{
market_capping(mv, capped_wts)
}
\arguments{
\item{mv}{Vector of market values}

\item{capped_wts}{Vector of weights to cap each market values}
}
\value{
A datafrane of results containing capped market values, capped weights, scaling factors and rescaled scaling factors (see Note)
}
\description{
Market_capping
}
\note{
When a country has zero market value, its scaling factor will be assigned 0, as opposed to 1 per the scaling factor spreadsheet convention, as it would give a more "correct" rescaled scaling factor.

Rescaled scaling factors ensures that the largest factor is 1 by rescaling scaling factors.
}
\examples{
data(mvs)
data(caps)
market_capping(mvs, caps)
}
