% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat.R
\name{pad_2zeros}
\alias{pad_2zeros}
\title{Pad numbers with two zeros}
\usage{
pad_2zeros(Number)
}
\arguments{
\item{Number}{to be padded}
}
\value{
Number padded with two zeros
}
\description{
Pad numbers with two zeros
}
\details{
Helper function for using dates with Rblpapi
}
