% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset_cors.R
\name{plot_asset_cors}
\alias{plot_asset_cors}
\title{Plot asset class correlations}
\usage{
plot_asset_cors()
}
\value{
Correlation plot
}
\description{
Downloads data from Bloomberg on major asset classes and plots their correlation
}
\examples{
\donttest{plot_asset_cors}
}
