% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat.R
\name{plot_credit_ratings}
\alias{plot_credit_ratings}
\title{Plot credit ratings evolution over time and facet by country}
\usage{
plot_credit_ratings(my_data)
}
\arguments{
\item{my_data}{The data framme generated by get_bm_ratings()}
}
\value{
Returns a plot
}
\description{
Plot credit ratings evolution over time and facet by country
}
\examples{
\donttest{get_bm_ratings(start_date="2018-10-01", end_date = "2018-11-01") \%>\% plot_credit_ratings()}
}
