% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_charts.R
\name{plot_cycles}
\alias{plot_cycles}
\title{Plot cycles chart}
\usage{
plot_cycles(df, cycles, start_date = NA, title = "",
  xlab = "Proportion of cycle", ylab = "")
}
\arguments{
\item{df}{dataframe containing data and value}

\item{cycles}{dataframe containing cycles with columns `from` and `to` as Dates}

\item{start_date}{required start date if any}

\item{title}{title of the chart}

\item{xlab}{xlab for the chart}

\item{ylab}{ylab for the chart}
}
\description{
Plot cycles chart
}
\examples{
emu_cycles <- data.frame(from = as.Date(c("1993-07-01", "2009-04-01", "2013-01-01")), to = c(as.Date(c("2008-04-1", "2011-10-01", "2018-12-31"))))
          df <- getData(ticker, fields = "PE_RATIO")
          plot_cycles(df, emu_cycles,
}
