% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_charts.R
\name{plot_cycles_bbg}
\alias{plot_cycles_bbg}
\title{Plot cycles chart using Bloomberg data}
\usage{
plot_cycles_bbg(ticker, field, cycles, start_date = NA, title = "",
  xlab = "Proportion of cycle", ylab = "")
}
\arguments{
\item{ticker}{BBG ticker of securty}

\item{field}{BBG field code of the security}

\item{cycles}{dataframe containing cycles with columns `from` and `to` as Dates}

\item{start_date}{required start date if any}

\item{title}{title of the chart}

\item{xlab}{xlab for the chart}

\item{ylab}{ylab for the chart}
}
\description{
Plot cycles chart using Bloomberg data
}
\examples{
plot_cycles_bbg("MXEM Index", "EQY_DVD_YLD_12M", cycles,
}
