% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charting.R
\name{scale_color_strat}
\alias{scale_color_strat}
\title{Colour schemes for ggplot}
\usage{
scale_color_strat()
}
\description{
Colour schemes for ggplot
}
\examples{
library(ggplot2)
data(pe_ratio_mxem)
df <-  dplyr::mutate(pe_ratio_mxem, date_range = cut(date, 5))
ggplot(df, aes(x = date, y = PE_RATIO, color = date_range)) + geom_line() + scale_color_strat()
}
