% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat.R
\name{show_bm_ratings_plot}
\alias{show_bm_ratings_plot}
\title{Show median credit rating over time}
\usage{
show_bm_ratings_plot(start_date = Sys.Date() - 365,
  end_date = Sys.Date(), freq = "1 month", date_adjust = 1)
}
\arguments{
\item{start_date}{Start date to show credit ratings.}

\item{end_date}{End date to show credit ratings.}

\item{freq}{Frequency of data.}

\item{date_adjust}{Date adjustment parameter: included as an easy way to generate monthends. Use start and end dates as the first day of each month, and set date_adjust=1 to subtract from the sequence of dates. This gives us the monthends.}
}
\value{
Generates a plot to show evolution of credit rating for benchmark countries
}
\description{
Show median credit rating over time
}
\examples{
##show_bm_ratings_plot()
}
