% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charting.R
\name{theme_strat}
\alias{theme_strat}
\title{Charting theme for ggplot}
\usage{
theme_strat()
}
\description{
Charting theme for ggplot
}
\examples{
library(ggplot2)
data(pe_ratio_mxem)
ggplot(pe_ratio_mxem, aes(x = date, y = PE_RATIO)) + geom_line() + theme_strat()
}
