% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{ts_augment}
\alias{ts_augment}
\title{Augment a time series with new features}
\usage{
ts_augment(data, date = date, price = .data$PX_LAST,
  rf_rate = 0.04/252)
}
\arguments{
\item{data}{Dataframe containing data}

\item{date}{Column containing date information. Defaults to "date".}

\item{price}{Column containing price information. Defailts to "PX_LAST".}

\item{rf_rate}{Risk-free rate for calculating excess returns}
}
\value{
A dataframe containing extra features.
}
\description{
Augments a time series data frame with quantitative calculations
}
