% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio.R
\name{z_score}
\alias{z_score}
\title{Calculate Z-score}
\usage{
z_score(my_ts)
}
\arguments{
\item{my_ts}{Time series to calculate z-score}
}
\value{
Z-scores
}
\description{
Calculate Z-score
}
\examples{
z_score(1:10)
#[1] -1.4863011 -1.1560120 -0.8257228 -0.4954337 -0.1651446  0.1651446  0.4954337  0.8257228  1.1560120  1.4863011
}
