% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat_active_risk.R
\name{build_strategies}
\alias{build_strategies}
\title{Build strategies from an input csv file or dataframe of strategies, can use `demo_strategies` to generate a template file}
\usage{
build_strategies(input, start_date = as.Date("2000-01-01"),
  end_date = today())
}
\arguments{
\item{input}{A dataframe containing strategy inputs or A character variable containing the path to the input csv file}

\item{start_date}{A Date variable with start date of generation, this can be earlier or later than the dates in the template file}

\item{end_date}{A Date variable containing the end date for the position sizes, defaults to today}
}
\value{
A list containing the actual sizes (mix of months-weighted and percentage based on input file) and simulated sizes ie actual sizes but filled upwards, then downwards to use for simulation of returns
}
\description{
Build strategies from an input csv file or dataframe of strategies, can use `demo_strategies` to generate a template file
}
\examples{
data(demo_strategies)
\donttest{build_strategies(demo_strategies, as.Date("2016-01-01"), as.Date("2018-12-07"))}
\donttest{
build_strategies("input_file.csv")
}
}
