% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat_active_risk.R
\name{calc_active_risk}
\alias{calc_active_risk}
\title{Calculate active risk given return of strategies and weight of the strategies}
\usage{
calc_active_risk(unwt_ret_w_date, curr_wt, start_date = today() -
  years(10), end_date = today(), annualize_factor = 250)
}
\arguments{
\item{unwt_ret_w_date}{dataframe of returns of strategies}

\item{curr_wt}{dataframe (one row) or vector of weights, df must have colnames and vector must have names corresponding to colname of returns}

\item{start_date}{start_date for calc of active risk}

\item{end_date}{end_date}

\item{annualize_factor}{factor to multiply by to convert to annual. 250 for daily data (default) and 12 for monthly data}
}
\value{
list containing `cov_matrix`, `port_sd` (portfolio std dev), `active_risk`, `marginal_risk`
}
\description{
Calculate active risk given return of strategies and weight of the strategies
}
\examples{
unwt_ret_w_date <- data.frame(date = as.Date(c("2018-01-02", "2018-01-02", "2018-01-04", "2018-01-05")),
                              long_spx = c(0.015, 0.021, -0.03, 0.01),
                              long_ukx = c(-0.005, 0.03, -0.01, -0.04))
curr_wt <- data.frame(long_spx = 0.01, )
results <- calc_active_risk(unwt_ret_w_date, curr_wt)
results$active_risk
}
