% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat_active_risk.R
\name{calc_strat_headline_size}
\alias{calc_strat_headline_size}
\title{Calculates headline sizes of strategies given the size of the component trades by first summing the positive sizes and negative sizes separately. Then taking the higher number as the size}
\usage{
calc_strat_headline_size(strat_df)
}
\arguments{
\item{strat_df}{A dataframe of strategies which are dataframes containing timeseries sizes of their component instruments. Columns should be `date`, `strategy`, `instrument`, `size`}
}
\value{
A dataframe containing the size of the strategies daily
}
\description{
Calculates headline sizes of strategies given the size of the component trades by first summing the positive sizes and negative sizes separately. Then taking the higher number as the size
}
\examples{
portfolios <- build_strategies(demo_strategies, as.Date("2016-01-01"))
actual_pf_size <- convert_dur_size(portfolios$actual, portfolios$summary, demo_duration)
calc_strat_headline_size(actual_pf_size)
}
