% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat.R
\name{get_bm_ratings}
\alias{get_bm_ratings}
\title{Download and calculate median credit rating over time}
\usage{
get_bm_ratings(end_date = Sys.Date(), per = 120)
}
\arguments{
\item{end_date}{Last date to show credit ratings.}

\item{per}{No. of historical monthends to retrieve}
}
\value{
Returns a data frame to be used with plot_credit_ratings()
}
\description{
Download and calculate median credit rating over time
}
\details{
This function rounds the end_date to the most recent monthend (in the past). Credit rating agencies
might re-rate countries anytime within a month, and in making this function we are standarding
the credit ratings of countries as at each monthend.
}
\examples{
\donttest{
library(Rblpapi)
blpConnect()

get_bm_ratings(end_date="2018-10-01", per=2)
}
}
