% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat.R
\name{get_bm_ratings}
\alias{get_bm_ratings}
\title{Download and calculate median credit rating over time}
\usage{
get_bm_ratings(start_date = Sys.Date() - 365, end_date = Sys.Date(),
  freq = "1 month", date_adjust = 1)
}
\arguments{
\item{start_date}{Start date to show credit ratings.}

\item{end_date}{End date to show credit ratings.}

\item{freq}{Frequency of data.}

\item{date_adjust}{Date adjustment parameter: included as an easy way to generate monthends. Use start and end dates as the first day of each month, and set date_adjust=1 to subtract from the sequence of dates. This gives us the monthends.}
}
\value{
Returns a data frame to be used with plot_credit_ratings()
}
\description{
Download and calculate median credit rating over time
}
\examples{
\donttest{get_bm_ratings(start_date="2018-10-01", end_date = "2018-11-01")}
}
