% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_battery.R
\name{load_bbg_data}
\alias{load_bbg_data}
\title{Load BBG Data}
\usage{
load_bbg_data(dbname = "mydatabase.db")
}
\value{
Loads sqlite data with data
}
\description{
Load BBG Data
}
\examples{
load_bbg_data()
}
