% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat.R
\name{market_capping}
\alias{market_capping}
\title{Market_capping}
\usage{
market_capping(mv, capped_wts)
}
\arguments{
\item{mv}{Vector of market values}

\item{capped_wts}{Vector of weights to cap each market values}
}
\value{
A vector of results containing capped market values, capped weights, scaling factors and rescaled scaling factors (see Note)
}
\description{
Market_capping
}
\examples{
data(mvs)
data(caps)
market_capping(mvs, caps)
}
