% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat_active_risk.R
\name{pf_summary}
\alias{pf_summary}
\title{Gets a sizes of each strategy in a portfolio at a specific date. Use to retrieve sizes in month weight terms}
\usage{
pf_summary(portfolio, as_of_date = NULL, approx = TRUE)
}
\arguments{
\item{portfolio}{A dataframe of strategies which are timeseries sizes of their component instruments, contains columns `date`, `strategy`}

\item{as_of_date}{A Date variable}

\item{approx}{A boolean indicating if `as_of_date` is unavailable, the next available earlier date should be used}
}
\value{
A dataframe containing the sizes of the strategies
}
\description{
Gets a sizes of each strategy in a portfolio at a specific date. Use to retrieve sizes in month weight terms
}
