% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charting.R
\name{plot_cycles}
\alias{plot_cycles}
\title{Plot cycles chart}
\usage{
plot_cycles(df, cycles_df, title = "", xlab = "Proportion of cycle",
  ylab = NULL)
}
\arguments{
\item{df}{dataframe containing date and value}

\item{cycles_df}{dataframe containing cycles with columns `from` and `to` as Dates}

\item{title}{optional title of the chart}

\item{xlab}{optional xlab for the chart}

\item{ylab}{optional ylab for the chart}
}
\description{
Plot cycles chart
}
\examples{
data(ez_biz_cycles)
data(pe_ratio_mxem)
plot_cycles(pe_ratio_mxem, ez_biz_cycles)
}
