% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat_active_risk.R
\name{simulate_history}
\alias{simulate_history}
\title{Calculate the returns in history given a set of static strategy weights}
\usage{
simulate_history(unwt_ret_w_date, curr_wt, start_date, end_date)
}
\arguments{
\item{unwt_ret_w_date}{A dataframe containing the timeseries unweighted return of strategies}

\item{curr_wt}{A dataframe containing one row of weights of strategies with strategy names as column names}

\item{start_date}{A Date variable indicating the start period of simulation}

\item{end_date}{A Date variable indicating the end period of simulation}
}
\value{
A dataframe containing the timeseries weighted return of the period
}
\description{
Calculate the returns in history given a set of static strategy weights
}
